+++ js/data.js
@@ -104,12 +104,4 @@ const createPhotos = () => {
   };
 };
 
-const photosArr = Array.from({length: MAX_VALUE_PHOTOS}, (_, index) => {
-  const photo = createPhotos();
-  photo.id = index + 1;
-  photo.url = `photos/${photo.id}.jpg`;
-  return photo;
-});
-
-// eslint-disable-next-line no-console
-console.log(photosArr);
+export {createPhotos, MAX_VALUE_PHOTOS};
